+++ internal/delivery/orders/dto.go
@@ -19,7 +19,7 @@ type OrderResponse struct {
 	ID           uuid.UUID    `json:"id"`
 	OrderDate    time.Time    `json:"order_date"`
 	DeliveryDate time.Time    `json:"delivery_date"`
-	TotalPrice   uint16       `json:"total_price,omitempty"`
+	TotalPrice   uint32       `json:"total_price,omitempty"`
 	Address      string       `json:"address"`
 	Status       string       `json:"status,omitempty"`
 	Products     []ProductDTO `json:"products"`
@@ -29,7 +29,7 @@ type OrderMaxResponse struct {
 	ID           uuid.UUID    `json:"id"`
 	OrderDate    time.Time    `json:"order_date"`
 	DeliveryDate time.Time    `json:"delivery_date"`
-	TotalPrice   uint16       `json:"total_price,omitempty"`
+	TotalPrice   uint32       `json:"total_price,omitempty"`
 	Address      string       `json:"address"`
 	Status       string       `json:"status,omitempty"`
 	Recipient    string       `json:"recipient"`
@@ -38,8 +38,8 @@ type OrderMaxResponse struct {
 
 type ProductDTO struct {
 	ID       uint32  `json:"id"`
-	Cost     uint16  `json:"cost,omitempty"`
-	Count    uint16  `json:"count,omitempty"`
+	Cost     uint32  `json:"cost,omitempty"`
+	Count    uint32  `json:"count,omitempty"`
 	Weight   float32 `json:"weight,omitempty"`
 	ImageURL string  `json:"image_url"`
 	Name     string  `json:"name"`
@@ -53,7 +53,7 @@ func ToOrderResponse(o *order.Order) OrderResponse {
 			ID:       p.ProductID,
 			ImageURL: p.ImageUrl,
 			Name:     p.Name,
-			Cost:     p.Cost,
+			Cost:     p.Cost, // Приведение типа
 			Count:    p.Count,
 			Weight:   p.Weight,
 		})
