+++ internal/model/cart_for_order.go
@@ -1,7 +1,7 @@
 package model
 
 type CartForOrder struct {
-	TotalItems             uint16
+	TotalItems             uint32
 	TotalWeight            float32
 	FinalPrice             float32
 	DeliveryDates          []DeliveryDateForOrder
