+++ internal/model/order_from_cart.go
@@ -9,7 +9,7 @@ type OrderFromCart struct {
 	OrderID      uuid.UUID
 	UserID       uint32
 	Address      string
-	TotalPrice   uint16
+	TotalPrice   uint32
 	DeliveryDate time.Time
 	CreatedAt    time.Time
 	Products     []ProductOrder
